backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.46

# Configuration
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Logging
loguru==0.7.3

# AI / LLM
langgraph==0.2.60
langchain-openai==0.3.0
langchain-google-genai==2.0.8

# PDF generation
reportlab==4.2.5

# HTTP client
httpx==0.28.1

# Security
bcrypt==4.2.1
PyJWT==2.9.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0

# Linting
ruff==0.8.6
